[rename] resolver 디렉토리 생성 (#45)
